Updates to the following components: adobe-acrobat-sign, adobe-analytics, adobe-commerce-magento, adobe-io-events, adp-workforce-now, airtable, algolia, amazon-seller-central, amqp, anthropic, arcgis, arena-plm, asana, aspose, atlassian-jira, aws-dynamodb, aws-glue, aws-lambda, aws-redshift, aws-s3, aws-ses, aws-sns, aws-sqs, azure-blob, azure-cosmos-db, azure-event-grid, azure-files, azure-openai-service, azure-service-bus, bamboohr, bigcommerce, bill, box, bynder, calendly, canny, click-up, confluence, contentful, customer-io, databricks, datadog, deepseek, docusign, domo, dropbox, duro-plm, expensify, facebook-marketing, firebase, first-resonance, fluent-commerce, freshservice, frontify, ftp, github, gong, google-ads, google-analytics, google-analytics-ga4, google-calendar, google-cloud-bigquery, google-cloud-pub-sub, google-cloud-storage, google-content-shopping, google-docs, google-drive, google-gemini, google-gmail, google-sheets, gorgias, gotowebinar, greenhouse, guru, gusto, hibob, hubspot, imap, intercom, jira-service-management, kafka, karbon, klaviyo, ldap, mailchimp, marketo, mixpanel, monday, mongo, mqtt, ms-bing-ads, ms-bot-framework, ms-business-central, ms-dynamics, ms-entra-id, ms-excel, ms-graph-api, ms-intune, ms-onedrive, ms-outlook, ms-power-bi, ms-project, ms-sharepoint, ms-sql-server, ms-teams, mysql, netsuite, new-relic, notion, odoo, okta, openai, oracledb, pagerduty, paylocity, pdq, pipedrive, postgres, postmark, qlik, quickbooks, quickbooks-time, ramp, redis, rippling, sage, sage-200, sage-hr, sage-intacct, salesforce, salesforce-marketing-cloud, sap-business-one, sap-ecc, sap-successfactors, sapS4Hana, segment, sendgrid, servicedesk-plus, servicenow, servicetitan, sftp, shipbob, shipstation, shopify, slack, smartsheet, smtp, snowflake, square, stripe, surveymonkey, system-surveyor, tableau, teamviewer, tenable-vulnerability-management, toast, twilio, typeform, ukg-pro, whatsapp, woo-commerce, workday, xai-grok, xero, yoti-sign, zendesk, zendesk-sell, zoho, zoom. Updates to the following packages: atlassian-utils, aws-utils, ms-utils. (b38a29f)

+++ components/domo/src/actions/groups/removeUserFromGroup.ts
@@ -2,7 +2,6 @@ import { action } from "@prismatic-io/spectral";
 import { getDomoClient } from "../../client";
 import { removeUserFromGroupInputs } from "../../inputs";
 import { removeUserFromGroupExamplePayload } from "../../examplePayloads";
-
 export const removeUserFromGroup = action({
   display: {
     label: "Remove User From Group",
@@ -11,12 +10,9 @@ export const removeUserFromGroup = action({
   examplePayload: removeUserFromGroupExamplePayload,
   perform: async (context, { connection, groupId, userId }) => {
     const client = await getDomoClient(connection, context.debug.enabled);
-    const { data } = await client.delete(
-      `/groups/${groupId}/users/${userId}`,
-    );
+    const { data } = await client.delete(`/groups/${groupId}/users/${userId}`);
     return { data };
   },
   inputs: removeUserFromGroupInputs,
 });
-
 export default { removeUserFromGroup };
